start working in Add DATA SOURCE from Woocommerce , Elementor , ... v2

+++ database/migrations/2022_12_12_175809_add_platform_to_sources_table.php
@@ -13,9 +13,13 @@ class AddPlatformToSourcesTable extends Migration
      */
     public function up()
     {
-        Schema::table('sources', function (Blueprint $table) {
-            $table->string('platform')->after('name');
-        });
+
+        if (!Schema::hasColumn('sources', 'platform')) {
+
+            Schema::table('sources', function (Blueprint $table) {
+                $table->string('platform')->after('name');
+            });
+        }
     }
 
     /**
@@ -25,8 +29,11 @@ class AddPlatformToSourcesTable extends Migration
      */
     public function down()
     {
-        Schema::table('sources', function (Blueprint $table) {
-            $table->dropColumn('platform');
-        });
+        if (!Schema::hasColumn('sources', 'platform')) {
+
+            Schema::table('sources', function (Blueprint $table) {
+                $table->dropColumn('platform');
+            });
+        }
     }
 }
